// src/enums/logs.enums.ts
export enum AmazonLevelEnum {
  CAMPAIGN_LEVEL = 'Campaign',
  AD_GROUP_LEVEL = 'AdGroup',
  PRODUCT_ADS = 'ProductAds',
  TARGETING_LEVEL = 'Targeting',
  KEYWORD_TARGETING = 'Keyword-Targeting',
  NEGATIVE_KEYWORD_TARGETING = 'Negative-Keyword-Targeting',
  NEGATIVE_PRODUCT_TARGETING = 'Negative Product Targeting',
  PRODUCT_TARGETING = 'Product-Targeting',
  RULE = 'RULE',
}

export enum WalmartLevelEnum {
  CAMPAIGN = 'campaign',
  AD_GROUP = 'adGroup',
  KEYWORD = 'keyword',
  AD_ITEM = 'adItem',
  PAGE_TYPE = 'pageType',
  PLATFORM = 'platform',
  AUTO_SEARCH_TERM = 'auto-searchTerm',
  MANUAL_SEARCH_TERM = 'manual-searchTerm',
  BRAND_PROFILES = 'brandProfiles',
  RULE = 'RULE',
}
export enum LogsTitlesEnum {
  LOGS_HOME = 'Logs_Home_Page',
}

export enum ActionTypesEnum {
  MAX_BID = 'maxBid',
  MIN_BID = 'minBid',
  BUDGET = 'budget',
  NAME = 'name',
  TROAS = 'troas',
  STATE = 'state',
  STATUS = 'status',
  BID = 'bid',
  KEYWORD_BID = 'keywordBid',
  CAMPAIGN_NAME = 'campaignName',
  TOTAL_BUDGET = 'totalBudget',
  DAILY_BUDGET = 'dailyBudget',
  PLACEMENT_BIDDING = 'Placement Bidding',
  PLATFORM_TYPE = 'platformType',
  END_DATE = 'endDate',
  PLACEMENT_TYPE = 'placementType',
  STRATEGY = 'Strategy',
  TYPE = 'type',
}